Bài tập lớn trí tuệ nhân tạo

+++ src/Console2/PairAngleShape.java
@@ -8,5 +8,5 @@ import java.util.List;
 public class PairAngleShape {
     List<ShapeEntity> shapeEntities = new ArrayList<>();
     List<Integer> position = new ArrayList<>();
-    int totalAngle;
+    double totalAngle;
 }
