+++ src/Module/Shape/ShapeEntity.java
@@ -31,7 +31,7 @@ public class ShapeEntity implements Serializable {
     public int weight;
     public ShapeEntity combineA;
     public ShapeEntity combineB;
-    public List<Integer> angles;
+    public List<Double> angles;
     public double area;
     public ShapeEntity() {
     }
